Add reset calculator modal
